emailservice/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.24.0
google-auth==1.24.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
grpcio-health-checking==1.34.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry.instrumentation.grpc==0.26b1
opentelemetry-exporter-otlp-proto-http==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry--instrumentation-flask==0.26b1
opentelemetry--instrumentation-jinja2==0.26b1
opentelemetry--instrumentation-requests==0.26b1
opentelemetry--instrumentation-sqlite3==0.26b1
opentelemetry--instrumentation-urllib==0.26b1
protobuf==3.13.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-json-logger==2.0.1
pytz==2018.9
requests==2.24.0
rsa==4.0
six==1.15.0
thrift==0.13.0
uritemplate==3.0.0
urllib3==1.24.2
wrapt==1.12.1